Callback-based processing for bin/ruby-ls
