+++ solutions/part1/test/helper.rb
@@ -1,5 +1,5 @@
 require "minitest/autorun"
 
-require_relative "../lib/file_details"
+require_relative "../lib/ruby_ls"
 
 DATA_DIR = File.dirname(__FILE__) + "/../data"
